# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Uganda Wi-Fi Billing API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:4000/api/v1" }
  ],
  "paths": {
...
    "/orders/{id}": {
      "get": {
        "summary": "Get order status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}